// static-project/wsc-pc-trade/client/components/na/order/open-express-dialog/type.ts
export type IGetAlphaFeeRes = Omit<IGetDeliveryFeeRes, 'depositUrl'>;
// ...
export interface IPickTime {
  day: number | string;
  time: number;
}

export interface IExpressWayBill {
  expressWayBillType: number;
  expressId: number;
  auditNo: string;
  pickTime: IPickTime;
  expressFee: number;
  expressName: string;
  isPay: boolean;
  weight: number;
  printerId: number;
  printerKey?: string;
  printerDeviceNo?: string;
  printerChannel?: number;
  startAppointment?: number;
  endAppointment?: number;
}

export interface ICityChannelListItem {
  app_id: string;
  channel: number;
  is_cloud_tag: boolean;
  name: string;
  expired?: boolean;
}

export interface IDeliveryWindowItemDetailInfo {
  // 配送公司名称
  dist_company_name: string;
  // 配送单号
  dist_id: string;
  // 物流单号
  express_no: string;
  // 商品ID
  goods_id: number;
  // 商品数量
  num: number;
  // 单品多运其他包裹数量
  other_num: number;
  // 发货类型
  delivery_type: number;
  // 配送状态描述
  dist_status_desc: string;
  // 商品是否可选发货
  allow_check: boolean;
  // 商品发货状态描述
  delivery_status_desc: string;
  // 商品链接
  url: string;
  // 商品图片
  img_url: string;
  // 订单商品ID
  item_id: string;
  // 退款状态码  无退款中记录 no_refund  有退款中记录 had_process_refund  申请中退款都已经全部处理 all_done_refund
  refund_status_code: string;
  // 商品的配送状态
  dist_status: number;
  // 商品名称
  name: string;
  // 加料商品
  cy_ingredient_items: IDeliveryWindowItemDetailInfo[];
  // 配送公司编号
  dist_company_id: number;
  // 商品类目
  category: string;
  // 商品sku_iD
  sku_id: number;
  // 商品发货状态
  delivery_status?: number;
  weight: number;

  // 前端使用字段
  isLocalDelivery?: boolean;
  isSend: boolean;
}

export interface IGetDeliveryDetailRes {
  // 是否开启单品多运
  open_multiple_delivery?: boolean;
  // 是否开启智选发货
  alpha_express_enable?: boolean;
  // 自提信息
  self_fetch?: {
    fetch_address: string;
    fetch_name: string;
    fetch_time: string;
    user_tel: string;
    user_name: string;
  };
  // 配送方式描述
  dist_type_desc?: string;
  // 商品明细
  item_detail_info_list?: IDeliveryWindowItemDetailInfo[];
  // 是否整单发货
  is_whole_send?: boolean;
  // 无需发货
  no_need_to_deliver?: boolean;
  // 配送方式
  dist_type?: number;
  // 门店ID
  store_id?: number;
  // 是否开启同城配送
  open_city_delivery?: boolean;
  // 电子面单异常信息
  electronic_sheet_exception_info?: {
    electronic_sheet_exception_desc: string;
    electronic_sheet_exception_code: number;
  };
  // 开通的同城配送渠道
  city_channel_list: ICityChannelListItem[];
  // 周期购信息
  multi_period_delivery_info?: {
    delivery_time: string;
    last_estimate_delivery_time: string;
    period: number;
  };
  // 店铺余额，同城配送呼叫三方配送需要用到（单位：分）
  balance?: number;
  // 支持发货列表
  support_delivery_types?: string[];
  // 收货人信息
  consignee_info?: {
    consignee_name: string;
    consignee_tel: string;
    consignee_address: string;
  };
  // 待发货数量
  un_delivery_count?: number;
  delivery_count?: number;
}

export interface IModel extends IGetDeliveryDetailRes {
  hasRefund?: boolean;
}

interface IExpressItem {
  expressId?: number | '';
  expressNo?: string;
  expressName?: string;
}

export interface IItemPack {
  express: IExpressItem;
  num: number | string;
}

export interface ILocalDelivery {
  distWeight: number;
  tip: number | string;
  fee: number;
  totalFee: number | string;
  channel?: IGetAlphaFeeRes['distChannel'];
  appId?: string;
  isCloudTag: boolean;
  channelName?: string;
  feeStatus: number;
  depositUrl?: string;
}

export interface IDeliveryInfo {
  deliveryType: number;
  isSingleGoodsMultiExpress?: boolean;
  expressWayBill?: IExpressWayBill;
  express: IExpressItem;
  itemPackList?: IItemPack[];
  localDelivery?: Partial<ILocalDelivery>;
}